requirements.txt
beautifulsoup4
requests==2.10.0
django-cities-light
gunicorn>=19.7
Django==1.9.8
django-import-export
djangorestframework
simple-salesforce
phonenumbers
psycopg2